float to double in longFloat

+++ src/runtime/long.cpp
@@ -689,7 +689,7 @@ Box* longFloat(BoxedLong* v) {
     if (!isSubclass(v->cls, long_cls))
         raiseExcHelper(TypeError, "descriptor '__float__' requires a 'long' object but received a '%s'",
                        getTypeName(v));
-    float f = mpz_get_d(v->n);
+    double f = mpz_get_d(v->n);
     return new BoxedFloat(f);
 }
